requirements.txt
asgiref
colorama
dj-database-url
Django
Pillow
psycopg2
gunicorn
python-decouple
python-dotenv
reportlab
sqlparse
tqdm
typing_extensions
tzdata
crispy-bootstrap4==2022.1
whitenoise==6.4.0
oauth2client
django-debug-toolbar
django-cors-headers
djangorestframework
